# api/src/main/resources/static/swagger.yml
openapi: 3.0.3
info:
  title: Employee API
  version: 1.0.0
  description: Swagger spec for Employee Service
servers:
  - url: http://localhost:8111
    description: Local server
paths:
  /employees:
    get:
      summary: Get All Employees
      responses:
        '200':
          description: Success – list of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service Unavailable (Circuit Breaker)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewEmployeeRequest'
      responses:
        '201':
          description: Employee created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          description: Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service Unavailable (Circuit Breaker)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    Employee:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        salary:
          type: integer
        age:
          type: integer
        title:
          type: string
        email:
          type: string
    NewEmployeeRequest:
      type: object
      properties:
        name:
          type: string
        salary:
          type: integer
        age:
          type: integer
        title:
          type: string
      required:
        - name
        - salary
        - age
        - title
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        details:
          type: string
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            fieldErrors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string